pyproject.toml
[tool.poetry]
name = "qencore"
version = "0.1.0"
description = "Quantum implementation of Convex Regression and Optimal Transport"
authors = ["Juan C Gonzalez-Espitia <jues.carlos@gmail.com>", "Jannis Born <jab@zurich.ibm.com>"]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "3.8.0"
pandas = "2.1.1"
qiskit = "1.0.2"
qiskit_aer = "0.14.2"
scikit-learn = "1.3.0"
scipy = "1.13.1"
torch = "2.1.1"
typer = "^0.12.3"
absl-py = "^2.1.0"
pyyaml = "^6.0.1"
flake8 = "^7.1.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
isort = "5.13.2"
pytest-cov = "3.0.0"
black = "22.3.0"
mypy = "1.10.0"
ruff = "0.4.6"
flake8 = "^7.1.0"
pep8-naming = "^0.14.1"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pep8-naming = "^0.14.1"

[tool.mypy]
ignore_missing_imports = true
python_version = "3.11"

[[tool.mypy.overrides]]
module = "absl"
ignore_errors = true

